refactor: BaseActionHandler에 isLoading 추가
