+++ android/app/src/main/java/com/mulberry/ody/presentation/feature/meetings/model/MeetingsUiState.kt
@@ -1,11 +1,7 @@
 package com.mulberry.ody.presentation.feature.meetings.model
 
 sealed interface MeetingsUiState {
-    data object Loading : MeetingsUiState
-
     data object Empty : MeetingsUiState
-
-    data object Error: MeetingsUiState
 
     class Meetings(val content: List<MeetingUiModel>) : MeetingsUiState
 }
