+++ android/app/src/main/java/com/mulberry/ody/presentation/feature/setting/SettingScreen.kt
@@ -37,7 +37,7 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.compose.LifecycleEventEffect
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.mulberry.ody.R
-import com.mulberry.ody.presentation.common.ErrorSnackbarHandler
+import com.mulberry.ody.presentation.component.BaseActionHandler
 import com.mulberry.ody.presentation.common.modifier.noRippleClickable
 import com.mulberry.ody.presentation.component.OdySadDialog
 import com.mulberry.ody.presentation.component.OdyTopAppBar
@@ -126,7 +126,7 @@ fun SettingScreen(
         )
     }
 
-    ErrorSnackbarHandler(viewModel, snackbarHostState)
+    BaseActionHandler(viewModel, snackbarHostState)
     LifecycleEventEffect(event = Lifecycle.Event.ON_START) {
         viewModel.fetchNotificationSetting()
         settingState.updatePermission()
